+++ backend/main.py
@@ -1,6 +1,7 @@
 from fastapi import FastAPI
 
 from app.api.v1.routes.auth import router as auth_router
+from app.api.v1.routes.diagnosis import router as diagnosis_router
 from app.api.v1.routes.users import router as user_router
 from app.core.config import settings
 from app.core.logger import logger
@@ -16,6 +17,7 @@ app = FastAPI(
 
 app.include_router(auth_router, prefix="/auth", tags=["Auth"])
 app.include_router(user_router, prefix="/users", tags=["Users"])
+app.include_router(diagnosis_router)
 
 
 @app.on_event("startup")